norminetter sauf pour minimap
