+++ bonus/src/mlx_stuff_bonus.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   mlx_stuff_bonus.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: anhebert <anhebert@student.42.fr>          +#+  +:+       +#+        */
+/*   By: mraymond <mraymond@student.42quebec.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/30 14:16:13 by mraymond          #+#    #+#             */
-/*   Updated: 2022/12/01 12:16:50 by anhebert         ###   ########.fr       */
+/*   Updated: 2022/12/05 14:55:53 by mraymond         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,7 +35,6 @@ int	render_next_frame(t_vars *vars)
 			&vars->mlx.line_len, &vars->mlx.endian);
 	mlx_hook(vars->mlx.win, 2, 0, keypress_handler, vars);
 	raycast_main_loop(vars);
-	//minimap_init(vars);
 	draw_map(vars);
 	draw_perso(vars);
 	mlx_put_image_to_window(vars->mlx.mlx, vars->mlx.win,
