+++ bonus/src/mouse_move_bonus.c
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   mouse_move_bonus.c                                 :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: mraymond <mraymond@student.42quebec.com    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/12/05 14:56:09 by mraymond          #+#    #+#             */
+/*   Updated: 2022/12/05 14:56:14 by mraymond         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "../include/cub3d_bonus.h"
 
 int	mouse_move(int x, int y, t_vars *vars)
